+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.10] - 2019-05-08
 ### Added
 - Config to disable enchanted effect on gear items (this adds a new client config file). Since the effect is broken on vanilla models, this is disabled by default.
+- Spears. They are weaker than swords, but have a much longer reach. These may need some balancing still.
 ### Fixed
 - Should fix gear parts and traits not syncing to clients on LAN games